pyproject.toml
[project]
name = "test-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "langchain-community>=0.3",
    "python-dotenv>=1.0",
    "deepagents>=0.5.3",
    "langchain-google-genai>=4.2.2",
]

[project.scripts]
start = "main:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["test_module"]
